部署项目

+++ RST.sh
@@ -6,10 +6,14 @@ BLUE="\033[0;36m"
 green(){
     echo -e "\033[32m\033[01m$1\033[0m"
 }
-domainname=`sed -n "2p" /etc/v2ray/v2ray_info`
-id=`sed -n "1p" /etc/v2ray/v2ray_info`
-shadowsockspwd=`sed -n "1p" /etc/shadowsocks-r/ssr_info`
-domainname=`sed -n "5p" /etc/shadowsocks-r/ssr_info`
+if [[ -f "/etc/shadowsocks-r/ssr_info" ]]; then
+    shadowsockspwd=`sed -n "1p" /etc/shadowsocks-r/ssr_info`
+    domainname=`sed -n "5p" /etc/shadowsocks-r/ssr_info`
+fi
+if [[ -f "/etc/shadowsocks-r/ssr_info" ]]; then
+    domainname=`sed -n "2p" /etc/v2ray/v2ray_info`
+    id=`sed -n "1p" /etc/v2ray/v2ray_info`
+fi
 if [ -f "/usr/sbin/nginx" ]; then
     /usr/bin/certbot-2 renew
     sleep 2
@@ -25,7 +29,12 @@ end_times=$(date +%s -d "$end_time")
 now_time=$(date +%s -d "$(date | awk -F ' +'  '{print $2,$3,$6}')")
 RST=$(($((end_times-now_time))/(60*60*24)))
 if [ "${end_time}" != ""  ]; then
-    sed -i "/<li>证书有效期剩余天数/c <li>证书有效期剩余天数:${RST}</li>" /var/www/${shadowsockspwd}.html
-    sed -i "/<li>证书有效期剩余天数/c <li>证书有效期剩余天数:${RST}</li>" /var/www/${id}.html
+    if [[ -f "/etc/shadowsocks-r/ssr_info" ]]; then
+        sed -i "/<li>证书有效期剩余天数/c <li>证书有效期剩余天数:${RST}</li>" /var/www/${shadowsockspwd}.html
+    fi
+    if [[ -f "/etc/shadowsocks-r/ssr_info" ]]; then
+        sed -i "/<li>证书有效期剩余天数/c <li>证书有效期剩余天数:${RST}</li>" /var/www/${id}.html
+    fi
+
 fi
